Initial commit
Thid folder contains some methods to calculate the scroll geometry based on manufacturing and performance constraints. Additionally it contains files for simulating scroll compressor and post processing. It is the part of the electronic ennex for the work presenterd in Tanveer and Bradshaw (2021)
"Performance evaluation of low-GWP refrigerants in 1-100 ton scroll compressors"
presented in Int. J. of Ref.